+++ src/LocationPicker.stories.tsx
@@ -4,14 +4,17 @@ import {action} from '@storybook/addon-actions';
 
 export default {
   title: 'Location Selector',
+  component: LocationSelector,
   parameters: {
     layout: 'fullscreen',
   },
   args: {
     accessToken: import.meta.env.STORYBOOK_VITE_MAPBOX_ACCESS_TOKEN ?? '',
-    geocoderProps: {},
-    initialPosition: undefined,
-    mapProps: {},
+    initialLocation: {
+      lng: -0.11865250984845943,
+      lat: 51.49935412224127,
+    },
+    onLocationChange: action('onLocationChange'),
   },
 } as ComponentMeta<typeof LocationSelector>;