+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Breaking changes:
 
 - Renamed `initialPosition` prop to `initialView`
+- Allow string value for `lng` or `lat` for initial value
 
 New features: